Moved the signal processing portion of fastzip
